+++ Spatialiser/findGroundPointForPixel.cc
@@ -1,6 +1,6 @@
 #include "spatialiser.ih"
 
-Maybe<Vector3d> Spatialiser::findGroundPointForPixel(Vector2i const& pixel, Affine3d const& cameraAgentTransform, double const groundZ) const
+Maybe<Vector3d> Spatialiser::findGroundPointForPixel(Vector2d const& pixel, Affine3d const& cameraAgentTransform, double const groundZ) const
 {
   Vector3d direction = cameraAgentTransform.rotation() * d_cameraModel->directionForPixel(pixel);
 
@@ -9,7 +9,7 @@ Maybe<Vector3d> Spatialiser::findGroundPointForPixel(Vector2i const& pixel, Affi
   return Math::intersectRayWithGroundPlane(position, direction, groundZ);
 }
 
-Maybe<Vector3d> Spatialiser::findGroundPointForPixel(Vector2i const& pixel, double const groundZ) const
+Maybe<Vector3d> Spatialiser::findGroundPointForPixel(Vector2d const& pixel, double const groundZ) const
 {
   Affine3d cameraAgentTransform = AgentState::get<BodyState>()->getCameraAgentTransform();
