+++ Spatialiser/spatialiser.hh
@@ -21,9 +21,9 @@ namespace bold
      *
      * @param pixel the x/y pixel location
      */
-    Maybe<Eigen::Vector3d> findGroundPointForPixel(Eigen::Vector2i const& pixel, double const groundZ = 0) const;
+    Maybe<Eigen::Vector3d> findGroundPointForPixel(Eigen::Vector2d const& pixel, double const groundZ = 0) const;
 
-    Maybe<Eigen::Vector3d> findGroundPointForPixel(Eigen::Vector2i const& pixel, Eigen::Affine3d const& cameraAgentTransform, double const groundZ = 0) const;
+    Maybe<Eigen::Vector3d> findGroundPointForPixel(Eigen::Vector2d const& pixel, Eigen::Affine3d const& cameraAgentTransform, double const groundZ = 0) const;
 
     int findHorizonForColumn(int column);
     int findHorizonForColumn(int column, Eigen::Affine3d const& cameraAgentTransform);
