+++ StateObject/CameraFrameState/cameraframestate.hh
@@ -13,16 +13,16 @@ namespace bold
   class CameraFrameState : public StateObject
   {
   public:
-    CameraFrameState(Maybe<Eigen::Vector2f> ballObservation,
-                     std::vector<Eigen::Vector2f> goalObservations,
+    CameraFrameState(Maybe<Eigen::Vector2d> ballObservation,
+                     std::vector<Eigen::Vector2d> goalObservations,
                      std::vector<LineSegment2i> observedLineSegments)
     : d_ballObservation(ballObservation),
       d_goalObservations(goalObservations),
       d_observedLineSegments(observedLineSegments)
     {}
 
-    Maybe<Eigen::Vector2f> getBallObservation() const { return d_ballObservation; }
-    std::vector<Eigen::Vector2f> getGoalObservations() const { return d_goalObservations; }
+    Maybe<Eigen::Vector2d> getBallObservation() const { return d_ballObservation; }
+    std::vector<Eigen::Vector2d> getGoalObservations() const { return d_goalObservations; }
     std::vector<LineSegment2i> getObservedLineSegments() const { return d_observedLineSegments; }
 
     bool isBallVisible() const { return d_ballObservation.hasValue(); }
@@ -30,8 +30,8 @@ namespace bold
     virtual void writeJson(rapidjson::Writer<rapidjson::StringBuffer>& writer) const override;
 
   private:
-    Maybe<Eigen::Vector2f> d_ballObservation;
-    std::vector<Eigen::Vector2f> d_goalObservations;
+    Maybe<Eigen::Vector2d> d_ballObservation;
+    std::vector<Eigen::Vector2d> d_goalObservations;
     std::vector<LineSegment2i> d_observedLineSegments;
   };
 }
